ATR-23: Exception helpers refactored into one
