+++ src/Acuminator/Acuminator.Utils/Acuminator.Utils.csproj
@@ -23,6 +23,7 @@
     <DefineConstants>DEBUG;TRACE</DefineConstants>
     <ErrorReport>prompt</ErrorReport>
     <WarningLevel>4</WarningLevel>
+    <LangVersion>latest</LangVersion>
   </PropertyGroup>
   <PropertyGroup Condition=" '$(Configuration)|$(Platform)' == 'Release|AnyCPU' ">
     <DebugType>pdbonly</DebugType>
@@ -92,7 +93,6 @@
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="CommonUtils\ArrayExtensions.cs" />
     <Compile Include="CommonUtils\EnumerableExtensions.cs" />
-    <Compile Include="CommonUtils\ExceptionHelper.cs" />
     <Compile Include="CommonUtils\ConcurrentExtensions.cs" />
     <Compile Include="CommonUtils\EnumerableExtensions - Copy.cs" />
     <Compile Include="RoslynExtensions\ITypeSymbolExtensions.cs" />
